+++ js/utils/check-string-length.js
@@ -3,3 +3,5 @@ function checkStringLength (string, length) {
 }
 
 checkStringLength ();
+
+export {checkStringLength};
